Massive cleanup of redundant actions
