+++ car/models.py
@@ -1,11 +1,7 @@
 from django.db import models
-from django.db.models.deletion import CASCADE
-from django.db.models.fields import BooleanField
 from django.utils.text import slugify
 import random
 import string
-
-from numpy.lib.twodim_base import triu_indices
 
 
 def rand_slug():
@@ -31,14 +27,7 @@ class Project(models.Model):
 
 class Batch(models.Model):
     project_id = models.ForeignKey(Project, on_delete=models.CASCADE)
-    batch_tag = models.CharField(max_length=20)
-
-class MculeQuote(models.Model):
-    project_id = models.ForeignKey(Project, on_delete=models.CASCADE)
-    quoteid = models.CharField(max_length=255)
-    quoteurl = models.CharField(max_length=255)
-    quotecost = models.FloatField()
-    quotevaliduntil = models.CharField(max_length=255)
+    batch_tag = models.CharField(max_length=50)
 
 
 class Target(models.Model):
@@ -51,7 +40,7 @@ class Target(models.Model):
     status = models.BooleanField(default=True)
     smiles = models.CharField(max_length=255, db_index=True, null=True)
     image = models.FileField(upload_to="targetimages/", max_length=255)
-    name = models.CharField(max_length=255, db_index=True, unique=True)
+    name = models.CharField(max_length=255, db_index=True)
     targetmass = models.FloatField()
     targetmols = models.FloatField()
     unit = models.CharField(choices=Unit.choices, default=Unit.mg, max_length=10)
@@ -79,7 +68,7 @@ class Reaction(models.Model):
 
 class Product(models.Model):
     reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    name = models.CharField(max_length=255, unique=True)
+    name = models.CharField(max_length=255)
     smiles = models.CharField(max_length=255, db_index=True, null=True)
     image = models.FileField(upload_to="productimages/", max_length=255)
     mculeid = models.CharField(max_length=255, null=True)
@@ -97,8 +86,7 @@ class AnalyseAction(models.Model):
     method = models.CharField(choices=QCMethod.choices, default=QCMethod.LCMS, max_length=10)
 
 
-# Models to capture IBM actions
-class IBMAddAction(models.Model):
+class AddAction(models.Model):
     class Unit(models.TextChoices):
         mmol = "mmol"
         ml = "ml"
@@ -140,74 +128,7 @@ class IBMAddAction(models.Model):
     mculedeliverytime = models.IntegerField(null=True)
 
 
-class IBMCollectLayerAction(models.Model):
-    class Layer(models.TextChoices):
-        organic = "organic"
-        aqueous = "aqueous"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    layer = models.CharField(choices=Layer.choices, default=Layer.organic, max_length=15)
-
-
-class IBMConcentrateAction(models.Model):
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-
-
-class IBMDegasAction(models.Model):
-    class Unit(models.TextChoices):
-        seconds = "seconds"
-        minutes = "minutes"
-        hours = "hours"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    gas = models.CharField(max_length=100)
-    duration = models.FloatField(null=True)
-    durationunit = models.CharField(choices=Unit.choices, default=Unit.seconds, max_length=10)
-
-
-class IBMDrySolidAction(models.Model):
-    class TemperatureUnit(models.TextChoices):
-        degcel = "degC"
-        kelvin = "K"
-
-    class Unit(models.TextChoices):
-        seconds = "seconds"
-        minutes = "minutes"
-        hours = "hours"
-
-    class Atmosphere(models.TextChoices):
-        nitrogen = "nitrogen"
-        vacuum = "vacuum"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    temperature = models.IntegerField(null=True)
-    temperatureunit = models.CharField(
-        choices=TemperatureUnit.choices, default=TemperatureUnit.degcel, max_length=10
-    )
-
-    duration = models.FloatField(null=True)
-    durationunit = models.CharField(choices=Unit.choices, default=Unit.seconds, max_length=10)
-    atmosphere = models.CharField(
-        choices=Atmosphere.choices, default=Atmosphere.nitrogen, max_length=10
-    )
-
-
-class IBMDrySolutionAction(models.Model):
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    dryingagent = models.CharField(max_length=100)
-
-
-class IBMExtractAction(models.Model):
+class ExtractAction(models.Model):
     class Unit(models.TextChoices):
         ml = "ml"
         mmol = "mmol"
@@ -222,7 +143,7 @@ class IBMExtractAction(models.Model):
     numberofrepetitions = models.IntegerField(null=True)
 
 
-class IBMFilterAction(models.Model):
+class FilterAction(models.Model):
     class PhaseToKeep(models.TextChoices):
         filtrate = "filtrate"
         precipitate = "precipitate"
@@ -254,81 +175,7 @@ class IBMFilterAction(models.Model):
         max_length=10,
     )
 
-
-class IBMMakeSolutionAction(models.Model):
-    class Unit(models.TextChoices):
-        ml = "ml"
-        mmol = "mmol"
-        moleq = "moleq"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    solute = models.CharField(max_length=255)
-    solutesmiles = models.CharField(max_length=255, null=True)
-    soluteimage = models.FileField(upload_to="IBMmakesolnimages/", max_length=255, null=True)
-    solvent = models.CharField(max_length=255)
-    solventsmiles = models.CharField(max_length=255, null=True)
-    solventimage = models.FileField(upload_to="IBMmakesolnimages/", max_length=255, null=True)
-    solutequantity = models.FloatField()
-    solutequantityunit = models.CharField(choices=Unit.choices, default=Unit.ml, max_length=10)
-    solventquantity = models.FloatField()
-    solventquantityunit = models.CharField(choices=Unit.choices, default=Unit.ml, max_length=10)
-
-
-class IBMPartitionAction(models.Model):
-    class Unit(models.TextChoices):
-        ml = "ml"
-        mmol = "mmol"
-        moleq = "moleq"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-
-    firstpartitionsolvent = models.CharField(max_length=255)
-    secondpartitionsolvent = models.CharField(max_length=255)
-    firstpartitionsolventquantity = models.IntegerField()
-    secondpartitionsolventquantity = models.IntegerField()
-    firstpartitionsolventquantityunit = models.CharField(
-        choices=Unit.choices, default=Unit.ml, max_length=10
-    )
-    secondpartitionsolventquantityunit = models.CharField(
-        choices=Unit.choices, default=Unit.ml, max_length=10
-    )
-
-
-class IBMpHAction(models.Model):
-    class TemperatureUnit(models.TextChoices):
-        degcel = "degC"
-        kelvin = "K"
-
-    class Unit(models.TextChoices):
-        ml = "ml"
-        mmol = "mmol"
-        moleq = "moleq"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    material = models.CharField(max_length=100)
-    materialquantity = models.FloatField(null=True)
-    materialquantityunit = models.CharField(choices=Unit.choices, default=Unit.ml, max_length=10)
-    pH = models.FloatField(null=True)
-    dropwise = models.BooleanField(default=False)
-    temperature = models.FloatField(null=True)
-    temperatureunit = models.CharField(
-        choices=TemperatureUnit.choices, default=TemperatureUnit.degcel, max_length=10
-    )
-
-
-class IBMPhaseSeparationAction(models.Model):
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-
-
-class IBMQuenchAction(models.Model):
+class QuenchAction(models.Model):
     class TemperatureUnit(models.TextChoices):
         degcel = "degC"
         kelvin = "K"
@@ -352,32 +199,7 @@ class IBMQuenchAction(models.Model):
     )
 
 
-class IBMRefluxAction(models.Model):
-    class Unit(models.TextChoices):
-        seconds = "seconds"
-        minutes = "minutes"
-        hours = "hours"
-
-    class Speed(models.TextChoices):
-        gentle = "gentle"
-        normal = "normal"
-        vigorous = "vigorous"
-
-    class Atmosphere(models.TextChoices):
-        nitrogen = "nitrogen"
-        air = "air"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    duration = models.FloatField(null=True)
-    durationunit = models.CharField(choices=Unit.choices, default=Unit.seconds, max_length=10)
-    stirringspeed = models.CharField(choices=Speed.choices, default=Speed.normal, max_length=10)
-    deanstarkapparatus = models.BooleanField(default=False)
-    atmosphere = models.CharField(choices=Atmosphere.choices, default=Atmosphere.air, max_length=10)
-
-
-class IBMSetTemperatureAction(models.Model):
+class SetTemperatureAction(models.Model):
     class TemperatureUnit(models.TextChoices):
         degcel = "degC"
         kelvin = "K"
@@ -391,7 +213,7 @@ class IBMSetTemperatureAction(models.Model):
     )
 
 
-class IBMStirAction(models.Model):
+class StirAction(models.Model):
     class TemperatureUnit(models.TextChoices):
         degcel = "degC"
         kelvin = "K"
@@ -422,42 +244,13 @@ class IBMStirAction(models.Model):
     stirringspeed = models.CharField(choices=Speed.choices, default=Speed.normal, max_length=10)
     atmosphere = models.CharField(choices=Atmosphere.choices, default=Atmosphere.air, max_length=10)
 
-
-class IBMStoreAction(models.Model):
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    material = models.CharField(max_length=255)
-    materialsmiles = models.CharField(max_length=255, null=True)
-    materialimage = models.FileField(upload_to="storeimages/", max_length=255, null=True)
-
-
-class IBMWaitAction(models.Model):
-    class Unit(models.TextChoices):
-        seconds = "seconds"
-        minutes = "minutes"
-        hours = "hours"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    duration = models.FloatField(null=True)
-    durationunit = models.CharField(choices=Unit.choices, default=Unit.seconds, max_length=10)
-    temperature = models.IntegerField(null=True)
-
-
-class IBMWashAction(models.Model):
-    class Unit(models.TextChoices):
-        ml = "ml"
-
-    reaction_id = models.ForeignKey(Reaction, on_delete=models.CASCADE)
-    actiontype = models.CharField(max_length=100)
-    actionno = models.IntegerField()
-    material = models.CharField(max_length=255)
-    materialquantity = models.FloatField(null=True)
-    materialquantityunit = models.CharField(choices=Unit.choices, default=Unit.ml, max_length=10)
-    numberofrepetitions = models.IntegerField(null=True)
-
+# Mcule models
+class MculeQuote(models.Model):
+    project_id = models.ForeignKey(Project, on_delete=models.CASCADE)
+    quoteid = models.CharField(max_length=255)
+    quoteurl = models.CharField(max_length=255)
+    quotecost = models.FloatField()
+    quotevaliduntil = models.CharField(max_length=255)
 
 class MCuleOrder(models.Model):
     project_id = models.ForeignKey(Project, on_delete=models.CASCADE)
