+++ car/templates/backend/upload.html
@@ -26,7 +26,7 @@
                       Here, you can upload CAR project targets for follow up compounds to be reviewed and synthesised
                   </p>
                   <br>
-                  <p>The current specification version is <b>ver_1.0</b></p>
+                  <p>The current specification version is <b>ver_1.2</b></p>
                   <br>
                   <form method="post" enctype="multipart/form-data">
                       {% csrf_token %}
