+++ requirements.txt
@@ -130,4 +130,4 @@ pubchempy==1.0.4
 pycule==0.1.9b0 
 ratelimit==2.2.1
 rxn4chemistry==0.4.0
-
+black==21.12b0
